routes for parentfeed complete just need to add user info and category search

+++ server/routes/delete_comment.js
@@ -7,22 +7,35 @@
 
 const express = require('express');
 const router  = express.Router();
+const verifyToken = require("./helpers");
 
 module.exports = (db) => {
-  router.delete("/:id", (req, res) => {
-  //add cookie session for the user_id to attach to logged in user
+  router.delete("/:id", verifyToken, (req, res) => {
+    const user = req.user
     const id = req.params.id;
-    console.log(req.body)
-    db.query(`DELETE FROM comments
+    console.log(id)
+    db.query(`SELECT users.id as users_id
+    FROM comments
+    LEFT JOIN users ON comments.user_id = users.id
     WHERE comments.id = $1;`,[id])
-      .then(data => {
-        res.status(201).json({"message": "comment is deleted"})
-      })
-      .catch(err => {
-        res
-          .status(500)
-          .json({ error: err.message });
-      });
+    .then(data => {
+      console.log(data);
+      const commentUserId = data.rows[0].users_id;
+      if (commentUserId !== user.id){
+       return res.status(400).send({ status: "error", message: "You can only delete your own comment" });
+      }
+
+      db.query(`DELETE FROM comments
+      WHERE comments.id = $1;`,[id])
+        .then(data => {
+          res.status(201).json({"message": "comment is deleted"})
+        })
+        .catch(err => {
+          res
+            .status(500)
+            .json({ error: err.message });
+        });
+    });
   });
   return router;
 };
