+++ server/routes/delete_posts.js
@@ -7,21 +7,36 @@
 
 const express = require('express');
 const router  = express.Router();
+const verifyToken = require("./helpers");
+
 
 module.exports = (db) => {
-  router.delete("/:id", (req, res) => {
-  //add cookie session for the user_id to attach to logged in user
+  router.delete("/:id", verifyToken, (req, res) => {
+    const user = req.user
+    console.log(user)
     const id = req.params.id;
-    console.log(req.body)
-    db.query(`DELETE FROM posts
-    WHERE posts.id = $1;`,[id])
+
+    db.query(`SELECT users.id as users_id
+         FROM posts
+         LEFT JOIN users ON posts.user_id = users.id
+         WHERE posts.id = $1;`,[id])
       .then(data => {
-        res.status(201).json({"message": "post is deleted"})
-      })
-      .catch(err => {
-        res
-          .status(500)
-          .json({ error: err.message });
+        console.log(data.rows[0].users_id)
+
+        const postUserId = data.rows[0].users_id;
+         if (postUserId !== user.id){
+          return res.status(400).send({ status: "error", message: "You can only delete your own post" });
+         }
+        db.query(`DELETE FROM posts
+        WHERE posts.id = $1;`,[id])
+          .then(data => {
+            res.status(201).json({"message": "post is deleted"})
+          })
+          .catch(err => {
+            res
+              .status(500)
+              .json({ error: err.message });
+          });
       });
   });
   return router;
